DITFReader: explicitly support 2x and 3x wide lines
The ITF specs says the wide lines may be 2 or 3 times as wide as the narrow ones.
The old code hard-coded 3x and managed to work also for some 2x codes by allowing
a greater MAX_INDIVIDUAL_VARIANCE.

This patch makes all 14 samples in the itf-1 set pass in fast-mode (instead of
only 9 as before).