regression/fix in UPCEANReader
The switch to use the generic FindPattern helper introduced a
‘regression’: one more false positive was detected. A closer look at
the example showed that the change was no regression, it was actually a
fix:

The upstream JAVA project implemented the
UPCEANReader::FindStartGuardPattern like this: scan for a pattern 111
and look for a quite-zone of 3 after the fact. If there is no quite
zone, then skip the whole pattern und start over. This fails to
identify the valid start guard given the input: 49333...
(falsepositives-2/14.png line 471).

This commit implements a (disabled) variant of the old behavior for
reference/documentation.