DecoderResult cleanup
Cleanup construction of DecoderResult return values. Goal was to
replace the pattern

DecoderRestult result;
result.setA(…);
result.setB(…);
return result;

Since not all properties are required in all contexts, the possibility
to selectively initialize values needed to be preserved.

This was achieved by a simple macro to set up getter/setter methods
that save lots of boilerplate. It sets up a standard 'const & () const'
getter, 2 setters for setting lvalues via copy and 2 for setting
rvalues via move. They are provided each to work either on lvalues
(normal 'void (...)') or on rvalues (returning '*this' as rvalue). The
latter can be used to optionally initialize a temporary in a return
statement, e.g.

    return DecoderResult(bytes, text).setEcLevel(level);

The rvalue setters are used to move potentially large data (strings,
byte arrays, etc.) into the result instead of copying it.