DMECEncoder: cleanup and bugfix
There was a bug in the data matrix encoder that caused “long” messages
to be encoded into an invalid matrix. The function EncodeECC200 was
switched from return-value to in-out-param while porting from JAVA. The
usage of the function has been updated only on the ‘small’ messages
code path, not on the ‘large’ one. See new DMWriterTest::Large.

This changeset fixes the bug and simplifies the EncodeECC200 function
by replacing the elaborate string splicing/copying with a more concise
(and faster) in-place version.

Furthermore: put the polynomial factors in an array-of-vectors
structure and use a faster binary search for the lookup of the
correctly sized polynomial. And also switch from ints to bytes, which
is what upstream is doing and this saves another conversion copy.