Fix a porting bug in DMHighLevelEncoder
Fix a bug that could cause the 3rd last character in a string to get
lost when TEXT encoding is used. The std::function<> replacement of the
JAVA virtual function call was missing one instance that led to the
call of the C40::EncodeChar instead of Text::EncodeChar.

That was a rather tricky one ;). Sticking with a runtime polymorphism
class hierarchy might have been the better choice here.