DMHighLevelEncoder: fix EDIFACT end of data
More random data found another EDIFACT encoder bug: the string
`https://test~[******]_` results in a valid code but if decoded again
results in a different string: `https://test~[******]{_`.

In the very special case of an EDIFACT segment ending with only the 31
unlatching code in the EOD buffer and two more code words available and
one remaining character in the message, the decoder does not expect to
see the unlatching code (31) and interprets it as a `{` character.

Solution: don’t add the 31 inside the EDIFACT segment in that case.
(Who comes up with such a convoluted special handling???)