DM bit placement/parsing consolidation (III)
Finally also extract the common ‘visiting pattern’ code from the
placement and the parsing function. The new VisitMatrix function gets a
functor that is responsible for processing each symbol (corner or utah
shape) one after another. It gets passed a BitPosArray with the
coordinates of each individual bit inside the symbol.

This separates the two concerns of 1) zig-zag scanning through the
individual bits of the matrix and 2) doing the actual processing
(reading or writing) at each bit position.