DMDetector: fix deadlock in traceGaps
Fix #30: The ‘go outward’ part in traceStep could result in adding
points to the top-line that were ‘behind’ the left line, which lead to
a top-line that was  parallel to the left line, which in turn, could
cause the ‘pull the current position back out onto the line’-part to
effectively reverse the last traceStep progress -> endless loop.

This change introduces a check after each gap, to see if we are still
between the start- and the end-line.