DMDetector: use PointI for current position in tracer
At the beginning of the development of the new detector algorithm, the
current position of an EdgeTracer could be a floating point position.
It later turned out, after each step, the position needed to be
centered on the pixel. So the code was effectively rounding the values
already but the type suggested there could be fractional values. Fix
this to improve readability of the code.

There was in fact one spot where it has not been rounded, which was a
bug.