DMWriter: fix another EDIFACT encoder bug
Fix another issue with the unlatching inside the EDIFACT encoder:
https://github.com/zxing/zxing/issues/960.

The problem only occurred when the shape was not forced to either
rectangular or square which resulted in a change of the symbol shape
during encoding of the last 2 characters.

My confidence level that this fix is complete and does not introduce a
regression is < 100%. For the encoder part it might be a good idea to
simply wrap the zint c library, which looks way cleaner, supports more
codes and seems to be actively maintained, too.