fix <filesystem> detection in TestReader
The feature macro __cpp_lib_filesystem is declared inside the <filesystem> header, so that needs to be included before checking for the macro. Macro checking is still required because e.g. libstdc++ may provide <filesystem> but it will be empty if the -std is set to anything lower than c++17.