fix gcc 5.4 compile error (partly revert warning fix)
It turns out, (c++11) compilers seem to disagree wheather or not the local variable 'pattern' needs to be  captured in a lambda or not. Reintroduce (the potentially bogous) warning of clang-6 but fix the error of gcc-5.4.
I actually did not understand why clang claimed that capture was unnecessary in the first place. But gcc-8 was happy in both cases...