perf: improve master ODRowReadre::FindPattern
Instead of iterating bit-by-bit and single-step incrementing the
counter, use the getNextSetTo() method and set the counter at the end.

Together with the other 'perf' commits, this reduces the runtime for the
falsepositive tests from around 15s to below 10s. The number of executed
instructions is reduced by around 20%, the number of branches by 40%.