Remove 'external' state of ODRowReaders
The DecoderState instances that had to be passed through all decodeRow functions and were the only reason to have dynamic type casting in the library were removed and replaced by standard internal state where required (only in RSS14 and RSSExtended). This required to introduce a reset() function to reset the internal state. This needs to be called before starting to scan a new image.

Makes the code cleaner and is also more in line with upstream.