Clean up UPCA/EAN extension blackbox tests
There were 6 files in upca-1 that contain extension codes. Now that we can check for valid metadata from extions move them to a separate folder and add .metadata.txt. 2 of the 6 actually fail to decode the extension properly.

Rename upcaean-extension-1 to ean13-extension-1 as this folder allows only to contain ean13 files.