ODUPCEReader: Fix a severe EAN8/UPCE decoder issue
This replaces the "FindGuardPattern" function which was used to search for the middle and end patterns in a scanning manner, much like the FindStartGuardPattern, with a non-scanning ReadGuardPattern version.

The old code would sometimes accept EAN8/UPCE codes from valid EAN13/UPCA codes by detecting only 4 digits and then scanning forward to find the middle guard pattern, skipping over valid digits. This is wrong. The middle / end patterns have to come right after the last digit.

This fixes a number of misreads and most false positives.