Simplify blackbox testing code and switch it to C++-17
To run the blackbox tests, we now require a compliant c++17 compiler (mostly due to std::filesystem). This removes the now obsolete gdiplus "runner" which in turn allows to significantly simplify the generic runner as we don't need the abstracted image loading and file management anymore.

This fixes the build on an uptodate MinGW64 based windows setup with gcc-9. It will surely break outdated setups which used to work.