ReedSolomon: speed up of > 2x by replacing mod operator
While testing the impact of the recent simplification of the detection of
'pure' QRCodes, I noticed that about 60% of the whole decoding time is
spent in a one-line function. And most of that time in one '%' operation.

This patch makes the whole set of unit tests about twice as fast.

The `const int` is only for clarity. No speed improvements there.