Result: deprecate resultPoints and introduce position
All `SampleGrid` based 2D codes now use the same mechanism to calculate
the position of the barcode based on PerspectiveTransform that is is also
used for sampling the image.

The 1D codes now also report their position in the same format (rectangle
of height 1).

`resultPoints()` is still there for source compatibility but it returns
different contents now: the same as `position()`.

The UPC_EAN_EXTENSION is currently not covered in the position.

This pretty much implements #130.

This implements