Result: introduce Result::orientation(), deprecate Metadata::ORIENTATION
Providing the orientation of a symbol as metadata but the position as
Result member is inconsistent. The new orientation() is a convenience
function to retrieve a rounded int of the Position::orientation() value
which can be used a direct replacement for ResultMetadata::ORIENTATION.

Position::rotation has been renamed to 'orientation'.