WIP: Speedup RSS14/RSS_EXPANDED with PatternView based pre-filter
As a quick solution to the performance issue, this adds a filter to only
start the expensive (old) BitArray based detection in case there is
likely a finder pattern to be found. This brings already pretty much all
speed improvements that there may be (for the falsepositive test case),
which is around 25x.