PointT: make it consistent with the upstream definition of "pixel center"
The PointF helper class was used inside my new DMDetector in a way that
defined (0,0) to be the 'center' of the first image pixel. The upstream
code (e.g. in GridSampler) defined (0.5, 0.5) to be the center of the
first pixel. See also some new documentation in GridSampler.h

This changeset brings the DMDetector in line with the rest of the code
base and prepares a wider use of PointF in other detectors.

To properly convert a PointI to a PointF, use the centered() function.
To go the other way around, simply use the PointI() casting constructor.