QRDetector: tune algorithm to work with corrupt alignment pattern
The issue #1314 of the upstream project revealed a potential issue with
the localization of the bottom right alignment pattern. Now the code only
accepts the CenterOfRing result if that center pixel is black and the
1:1:1 pattern matches. Had to introduce a relaxed pattern matching for the
alignment pattern to not loose already successfully detected samples with
degenerated (small) black centers, e.g. qrcode-2/13.png.

Added a modified sample from the upstream issue to the blackbox tests.