DMDetector: improve subpixel correction of corner points
This is more of a theoretical improvement: For a non-square symbol moving
towards the opposite corner is not the same as moving along both adjacent
sides the same amount.

Also modifying each corner in turn slightly effects the correction of the
next corner based on the change of the previous corners.