BarcodeFormat: cleanup naming of constants
This provides new names for the enum values with consistent style compared
to other enums in the API. Meaning: replace the SCREAMING_SNAKE_CASE
notation with the CamelCase convention.

The rules applied were:
 1. Start with the "natural" name of the barcode format (like "QR Code")
 2. Remove '-' and ' ' characters

The ToString() function has been changed to return the "natural" name but
without space characters.

The BarcodeFormatFromString() function parses all former and presently
valid notation styles.

"RSS" has been changed to "DataBar" since that is its official name for a
couple of years now.

The old names are kept as [[deprecated]] aliases to keep it source
compatible with existing client code for the time being.

The python wrapper constants have not been kept source compatible as it is
not widely used, yet. If someone has a case for it, we can later add the
old names back.

The Java and WinRT wrappers have not been touched (lack of CI testability).