tests: remove about 35MB worth of black box samples images
We have 135MB sample images in our source tarballs. This annoys package
managers and is unreasonable. This is a first step at reducing this size
by simply removing a number of images that are very similar to other
ones such that they can be considered to be effectively redundant in
their testing coverage.