Code128: replace array<vector> with array<array>
This fixes a inconsistency between the reader and the writer code. The
reader had to check for the termination bar separately, while the writer
did include it in the digit data from the mentioned array.

This also reduces the code size of the lib by about 17kb, just by removing
the 100+ vector instantiations.