Aztec: quick fix for pure codes with asymmetric quite zone
This is a quick solution for the issue brought up in #128. The introduction
of the `isPure` flag into the Aztec detector has the only meaning to limit
the new (performance degrading) behavior to the special `isPure` use case.

This adds a blackbox test from the KDE project where the issue showed up.