QRCode: improve isPure detection
Make the 'pure' detector capable of handling symbols with a non-integer
module width. This can happen due to aliasing if the rendered symbol
width in pixels is not an integer multiple of the symbol width in modules.

This finds 4 more samples from the blackbox test set and also the example
from the KDE project mentioned in
https://github.com/nu-book/zxing-cpp/issues/128#issuecomment-730479329.