PDF417: fall back to non-isPure code in case of a checksum error
I gave up on the proper handling of pure codes where the module size is
between 1 < m < 2 (for now). The old code can handle this most of the time,
so we fallback to that in case we detected a code but got a checksum error.

The problem is that bars which are 1 module wide can end up rendered with
2 pixel width (instead of 1) due to aliasing. This screws up the relative
widths too much.