DataMatrix: fix a regression in RegressionLine class
The change 956b72e5e8dae4a49741ea84ea6b5bda3ff99aee was not 'good'. The
calculation of the gap sizes inside the regression line is dependent on
the that line to be cleaned of outliers. An example showing the issue is
dm-2.jpg.

This came up during the followup change to support mirrored symbols.