ResultMetadata: deprecate STRUCTURED_APPEND_* keys and provide replacement
Add structured append related properties to `Result` class and hook up the
existing QRCode and PDF417 implementation.

The names of the properties are up for debate if someone has a better
suggestion.

The `STRUCTURED_APPEND_*` values are still set a.t.m. to keep old client
code compiling and working for now but this will be removed in the next
release.