PDF417 Macro: fileId decode; segmentIndex 0 -> -1; Macro only symbol (#219)
Decodes the Macro PDF417 "fileId" codewords as 0-899 numbers, each 0-filled to width 3. Not doing any base 900 maths based on the example in ISO/IEC 15438:2015 Annex H.6, where (100, 200, 300) is transmitted as "100200300".

This is in contrast to the upstream project, where the `fileId` is treaded as text compaction content (which is wrong because it is not a bijective function).

Sets the default of segmentIndex in PDFDecoderResultExtra to -1, and checks before setting StructuredAppendInfo (though check not really needed).

Allows a symbol to consist solely of a Macro (i.e. removes check for resultString being empty checks for non-existence on resultString empty check), as stated in Annex H.2 note.

Unit test additions. Also adds test entry that pdf417-1/03-rot90.png is readable when rotated 90 degrees.

Also suppresses a GNC maybe-uninitialized warning.