Android: completely new implementation including demo app
This adds a new Android library `zxingcpp` and a demo app `ZXingCpp-Demo`.
Also added is a fully-fledged AndroidStudio project and grade build files
that use cmake as the native build tool for the c++ part.

Improvements with respect to the old wrapper:
 * cmake based
 * support for portrait and landscape image input
 * use the new `ReadBarcode` API internally
 * direct support for Androids `ImageProxy` class as input
 * Kotlin instead of Java ;)

The demo app has a few switches to play with the settings, including a
switch to use the upstream Java project for a quick performance comparison.

Disclaimer: this is a first version, the API of the wrapper library may
still change. As I'm a total Android newbie, suggestions for improvement
are welcome.

The old wrapper (jni and java folders and Android.mk) will later be
removed. This commit is related to #153.