GDI+: remove unmaintained wrapper
The GDI+ wrapper was originally created to support blackbox testing on
windows. That has long been superseded by the cross platform code. It
was not deemed worth it porting that code to the ReadBarcode API.

If a GDI+ wrapper to conveniently access `Gdiplus::BitmapData` is indeed
in demand, we can introduce a simple header only wrapper like the
`ZXingQtReader.h`.