Multi-Barcode: differentiate between multiple 1D-symbols of the same code
This implements a simple heuristic based on the relation between the
distance between a new line and an existing compared to the length of the
symbol.

This feature adds the benefit that 1D symbols now can have a "height" other
than 1 reported back in their position.

This detects now all 6 symbols from the 'second example' in #138.