Add Android library project
This makes it possible (and easy) to use ZXing C++ with any Android app.

Android components are distributed as so-called Android Archives (AAR)
which are basically JARs with Android resources. This commit adds a library
project that bundles the native libraries and the necessary (JNI) wrapper
in a generic form.

The resulting AAR file can easily be dropped in any Android app and makes
ZXing C++ immediately useable without having to clone and build the ZXing
C++ core libraries and JNI wrapper in the context of another app.

To check the fitness of the library, there's a sample/demo app beside
the library project. This app, and the library, is based on the sources
in `wrappers/android`.

However, the JNI wrapper of this library differs from `wrappers/android`
to make it usable for Android's legacy camera API, Camera2 as well as
CameraX. The returned result (object) now contains all available data
(including binary data and meta stuff). And there are functions for
barcode generation too.

There's also an (untested!) JitPack integration to automatically publish
this library on jitpack.io as soon as there's a Release on GitHub for
version "v1.2.1" (see "afterEvaluate" in zxingcpp/build.gradle). JitPack
should build this library as soon as it is requested - I'm just unsure
if JitPack will find the (correct) `build.gradle` because it's in
`wrappers/aar` and not in the root folder. Unfortunately, there's no
way to test this apart from trying it.