OneD: fix a bug in the isPure code path for the stacked DataBar codes
Stacked DataBar (Extended) symbology is actually a 2D not 1D input, hence,
my former isPure shortcut had to fail for those. (thanks to @gitlost for
implicitly bringing that to my attention ;)). So for those DataBar
symbols, we actually need to process more than one non-empty line.

This code now requires that the middle line of the image passes through
the pure 1D barcode, which is a reasonable assumption imho.