Added code to encapsulate finder patterns for micro QR codes.
FinderPattern is provided to store the position of the micro qr code finder pattern and allow multiple candidates to be combined together.  FinderPatternInfo contains the position of the actual finder pattern plus two "fake" pattern positions which can be used in combination to correct for any distortion of the micro qr code.
Added simple unit tests for finder pattern code.