multi-res: add hack to be able to call downscale code in ReadBarcode
This is currently used in the android and python wrapper.

See discussion in #282 why this is not generally enabled at the moment.
It basically comes down to the fact, that we can not have different
default values for `tryDownscale` based on whether we call the single
or multi symbol function.