Added detector for micro qr codes.
Introduced new detector class that will search an input image for a micro qr code.  It will correct for any rotation or skew and then sample the image to produce a BitMatrix containing the QR code.  If the input image is a pure barcode (no skew/rotation) and the detector is given the pure barcode hint then it will sample the input image to produce a BitMatrix.
Added unit tests for the Detector class.