Added error correction levels for micro qr codes.
The error correction levels are similar to QR codes but there are some differences which explains the new class and tests.  Three bits (not two) are used to identify the error correction settings.  The high correction level setting is not supported for micro QR codes and there is a mode where error detection and not correction is provided for micro qr codes version 1.  Created two factory functions to create ErrorCorrectionLevel for QR and micro QR.