Added format information decoding.
The format information for micro QR codes is different from the QR codes.  Briefly, three bits are used to specify the error correction level and 2 for the data mask as opposed to 2 for error correction level and 3 for the data mask for QR codes.  The XOR mask for micro QR codes is different as well.  Factory functions are used to create the format information so that a single FormatInformation class can be passed around.
Added unit tests that check all the different error correction levels.