AZDetector: switch float to double to fix surprising clang-13 regression
Switching from clang-12 to clang-13 resulted in a failure of the
samples/aztec-1/readerinit-compact-15x15.png test, due to a numerical
instability. Seemingly, older clang and also g++-11 and g++-12 internally
operate in double precision with the 'ratio' value even though it used to
be a float, while clang-13 does no anymore.