Add micro qr code detector (#324)
This adds support to detect Micro QR Codes. The code was originally based on an implementation ported from Java that was completely separate from the `QR*` files. It has been rewritten such that it merges with the existing `QRCode` functionality, thereby increasing its performance (both detection rate and runtime speed). For the sake of a cleaner git history, the PR has been squashed to prevent around 3k lines of code of going in and out again immediately. 

Co-authored-by: axxel <awagger@gmail.com>