Improve binary data support (1 of N)
This is part of a somewhat larger effort to improve the out of the box
support for reading barcode symbols that contain binary data. A lengthy
discussion about the goals and design decisions can be found in #334.

From the user facing API, this manifests in the new property
`ByteArray Result::binary()`. This is still a WIP and may change in the
future.