cmake: only install minimal set of header files
Fixes #168. This has been tested to work with the example applications.
This is a pretty sharp "deprecation" of a whole lot of internal API and
reduces the published interface to merely `ReadBarcode` and
`MultiFormatWriter`. It is obviously somewhat risky but I see no other
practical way of moving forward with limiting the API.