Remove errors from reader
Now that we are returning an array of results I had to think about how to handle possible errors when decoding the codes.
Since I don't know of any case where the user of the lib is interested in the specific error, I decided not to return the error bound to a result.
If a code cannot be read, I simply don't return a result at all.