Add api to work directly with CVPixelBuffer
This should improve performance when the pixel buffers pixel format is one of the supported
luminance based ones. In this case we just hand over the luminance plane directly to ZXing without any color space conversion.
For all other color spaces we keep on doing the color space conversion and call the CIImage api.