Adding a wrapper for iOS (#337)
Co-authored-by: @parallaxe 
Co-authored-by: Christian Braun braun.christian.bamberg@gmail.com

This pull request adds a wrapper of ZXing for iOS. It offers one class for reading and one for writing.
ZXing itself will be compiled to an universal framework, xcframework, that contains the binaries and
headers for the iOS-simulator and for iOS-devices. The wrapper-code is written in Objective-C++, whereas
the public part of the wrapper is Objective-C only. By this, it can also be used in Swift.

To build the xcframework, I had to add a section to the CMakeLists.txt. This enables building frameworks
for Apple platforms. Those frameworks are only either for the simulator, or for real devices. But they can be
combined to a single xcframework. This process is automated by the build-release.sh-script.
The section in the CMakeLists.txt should not change the behaviour on other platforms (tested on Linux so
far).

To use the wrapper, you first have to build the xcframework by calling the build-script and then adding
the Package.swift-file to your project. A demoproject is included, where also the usage is demonstrated.