DataBarExpanded: return standard conforming text instead of HRI
This is a breaking change. It fixes the problem that the DataBarExpanded
reader would implicitly return the "human readable interpretation" of
the content instead of the spec conforming string (see the section
"4. Scanned GS1 barcode data" in https://www.linkedin.com/pulse/gs1-application-identifier-syntax-dictionary-terry-burton/

A later change will allow to somehow specify that the HRI shall be
returned as text. You may use the existing function `HRIFromGS1()` from
`GS1.h` in the meantime.