ZX_FAST_BIT_STORAGE: remove defunct packed bits variant of BitArray/Matrix
The original bit-packed version of both BitArray and BitMatrix have not
been even compiling for a long time now. The code has been kept around
to not 'loose' the knowledge/work therein. If that should ever be of
importance again (e.g. in special low memory embedded scenarios), the git
history will have to be used to retrieve that information back.