ECI: replace Qt originated charset conversion code with libzueci
Big thanks to @gitlost for providing libzueci especially for this project
to be able to get rid of the questionable LGPL (with exception) code
copied from Qt (which copied it from other BSD licensensed places).

The libzueci code is fetched via cmake directly from its own git repo
during configuration. We may decide to include the code in this repo later
in case this fetch should be too much of a burden for package maintainsers.