Support reading inverted barcodes
Because inverted barcodes are fairly common (especially QR Codes)
and there should be a general option to read these. Otherwise, we
have to invert the input image in all clients what is not always
trivial or as performant as doing it in the library.

This commit adds a new hint "invert" which makes the threshold
algorithms return an inverted binary image so inverted codes can
be read. Please note this means *only* inverted barcodes can be
read when "invert" is set. It's not an also try-this type of hint.

Instead, the client is supposed to control when to flip an image,
because reading inverted barcodes doubles the processing time.
In mobile environments it may therefore be better to just flip every
second camera frame instead of trying to read positive and negative
codes in one frame.