Inversion: implement new `DecodeHints::tryInvert` feature
If `tryInvert` is enabled (the default) then symbologies that support it
(QRCode, DataMatrix and Aztec) are tested with black/white inversion.

This causes around 10% slowdown for the false positive use case if all
formats are searched for. The total slowdown of all blackbox tests is
about 5%.

In the WASM/iOS/WinRT wrapper the new flag is enabled together with the
existing `tryHarder` flag for now.

This fixes #172 and obsoletes #173.