DecodeHints: set default `TextMode` from `Plain` to `HRI`
Switch default mode from `Plain` to `HRI` (human readable interpretation).
This effectively reversed the 'breaking' change of the `DataBar` reader
by reintroducing the `HRI` format that was the only option until and
including v1.4. It therefore introduces other breaking changes, especially
with symbols containing non-text data.

This is related to the discussion in #338.