Add options to the example reader/writers to show usage information and exit
Call the option -help for consistency with other options,
but support the traditional GNU option style --help too.

Distros such as Debian are now shipping the examples in tools packages,
some more folks are being exposed to the example reader/writers.

This will assist folks who are used to an option existing to
find the usage information, which is usually behind --help.