QRDecoder: fix 1-bit error in second copy of FormatInfo in mirrored case
There was an issue with mishandling the 'Dark Module' (see Figure 25 in
spec) in the mirrored case, where this constant 1 was shadowing the true
bit at position 7. This decreased the error resilience in the mirrored
case when relying on the second copy of the FormatInfo. (Yes, this is a
very unlikely event).