android: fix buffer size for strings from the JVM
The string buffer must be specified in bytes, which can of course
differ from the number of characters in UTF-8 strings.

Using the number of characters from `GetStringLength` will result
in too small a buffer and eventually a buffer overflow if the Java
string contains UTF-8 encoded characters (consisting of two bytes).