AUTHORS: remove misleading/outdated author attribution
The AUTHORS.ZXing text file contained names of people involved in writing
the original Java code. While the zxing-cpp project is based on said Java
code, this file as the single source of authorship information is
misleading at best: it does not name any of the contributors that added
code to the c++ port, which is a substantial portion after all those years.