PDF417: new DecodeCodewords API
This cleans up the internal API somewhat, making it obvious that the
erasures array is not used anyway. Also preparing the introduction of
MicroPDF17 detection.

To give a meaningful and consistent `Result::ecLevel` for all PDF417
variants, that string now contains a percentage number of how many code
words have been used for error correction.