MultiFormatWriter: fix encoding binary std::string
While `MultiFormatWriter::encode(std::wstring)` can encode binary data
just fine, `MultiFormatWriter::encode(std::string)` corrupts the data
by trying to interpret it as UTF-8.

This may also break immediately because not all byte combinations are
allowed in UTF-8.

So when `_encoding` is set to `CharacterSet::BINARY`, the `std::string`
needs to be converted to a `std::wstring` of unsigned values, because
`TextEncoder::GetBytes(std::wstring)` calls `ToUtf8()` which cannot
handle negative values.

For example, without this commit, try:

	auto writer = MultiFormatWriter(BarcodeFormat::QRCode)
		.setEncoding(CharacterSet::BINARY).
	auto bitmap = writer.encode(std::string("\x7e\x7f\x80\x81"), 200, 200);

Which will result in: "ValueError: Unexpected charcode".

Please note ZXingWriter is NOT affected by this!
This works just fine:
	$ printf "\\x7e\\x7f\\x80\\x81" > file
	$ example/ZXingWriter -binary QRCode file out.png
Because:
1) it's calling `MultiFormatWriter::encode(std::wstring)` and
2) builds the `std::wstring` from `uint8_t`