Python: Fix wrapper build on MSVC not having __cplusplus for BitHacks
The `BitHacks.h` header needs the `__cplusplus` define to determine
whether to include `<bit>`. The MSVC compiler only defines this when it
is passed the `/Zc:__cplusplus` flag. The main library was already
setting this flag, but it not set when building the Python wrapper,
causing compilation to fail because `countr_zero` and `popcount` are
being used without `<bit>` being included.

Fixes #612