android: add consumerProguardFiles
To automatically add the necessary proguard configuration to
the consuming app of this library.

This way, there is no manual proguard configuration required
for the users of this library.

Note that this is only true if the library is included as an AAR
(or via maven), but not in a multi-module build (like it's the
case with the sample app), because `consumerProguardFiles` is not
used in multi-module builds.