ios: improve exception/error handling
Make `read()` throw so C++ exceptions can be caught in Swift.

So an app can't crash silently anymore when a C++ exception is thrown
while trying to read a barcode.

This also adds a function to properly set a NSError from a C++
exception to avoid code duplication in ZXIBarcodeReader/Writer.