Reduce: make consistent use of Reduce() vs. std::accumulate()
Thought about introducing std::reduce as a faster replacement
for std::accumulate. std::reduce() is not required to be a strict
left-fold operation, meaning the order of the op-application is not
sequenced left-to-right. This sounded like an optimization opportunity
but it turns out that for this use case it actually does not make a
difference (falsepositives runtime). And when tested with a large
std::vector<uint16_t> and proper auto-vectorization (e.g. clang++ -O2),
it turns out that std::accumulate() can be twice as fast as std::reduce.

There are 3 uses of std::accumulate (instead of Reduce) left in the
code base where a strict left-fold operation is required.