python: add support to write `bytes` as binary data
Adding support to write both `str` and `bytes` via `write_barcode`. If
passed a `bytes` buffer, it will be encoded as binary data. This can be
considered a breaking change, because passing a `bytes` buffer that
happened to contain a valid utf8 string was converted to a standard text
due to the automatic conversion support of pybind11.

This fixes #694.