BitHacks: fix random QRCode content on pre-Haswell Windows machines (#727)
Turns out `__lzcnt()` can function as a random number generator on old
Windows boxes. Nice.

This fixes #727 and supersedes #728, as it takes care to prevent even more
wrong results by calling it for non 32 bit values. A C++20 build would have
fixed this as well.

Thanks to @Abooow786 for the effort of debugging this.