PDF417: reduce binary size of library by 90kB
The single biggest symbol in the release build was the static constexpr
pre-computed `RatioTable` with 87kB of float data. Move that to heap
allocated memory, computed once per process on first use. There is simple
#ifdef to move it back to the old place, in case that should be preferred.

This is related to the move of the same data from .bss to .rodata here:
https://github.com/zxing-cpp/zxing-cpp/commit/bd7a9509075e9531261c65c180119a9c710034bf

Also remove another 5kb from .rodata by getting rid of 2700 times the
2 bytes value 0x0001 from the SYMBOL_TABLE.