Barcode: switch symbol() from BitMatrix to ImageView
The idea is to make the API smaller by not exposing the elaborate internal
BitMatrix class, just to get access to the new symbol. With the removal of
the MultiFormatWriter API, the BitMatrix header can also be removed from
the public API.

This also moves/copies the `ToString(BitMatrix)` functionality to the
new `WriteBarcodeToUtf8()` function.