aztec: Add support for detecting Aztec Runes
Runes are identical to the Core of a compact Aztec code, except that
their bits are alternately flipped. They contain exactly 8 bits of data.

We know that a compact Core is a Rune if:
a) it fails a normal Reed-Solomon decode AND
b) it passes when you flip every odd-numbered bit

Once we know this, we can shortcut most of the rest of the decode logic.