Add support for detecting and reading Aztec Runes (#763)
* aztec: Add support for detecting Aztec Runes

Runes are identical to the Core of a compact Aztec code, except that
their bits are alternately flipped. They contain exactly 8 bits of data.

We know that a compact Core is a Rune if:
a) it fails a normal Reed-Solomon decode AND
b) it passes when you flip every odd-numbered bit

Once we know this, we can shortcut most of the rest of the decode logic.

* aztec: Fix pure detector bug

LocateConcentricPattern starts at the middle and works out, but it still
expects the range of the entire pattern, not half of it. The only reason
this has worked so far is that non Rune Aztec codes are always at least
twice the size of the target pattern.

* aztec: Add capability to generate runes

---------

Co-authored-by: Morgan <morgan.harris@unsw.edu.au>