DMDecoder: keep on decoding if a format error occurs
This is inspired by the discussion in #774. The idea is to "get as much
information out of the barcode as possible", even if a format error (like
an invalid code word) or readerInit flag at the wrong position occurs. This
still results in an error condition but the user might still find value in
interpreting the (completely) decoded message.